requirements.txt
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
notebook
tavily-python
wikipedia
trustcall
langgraph-cli[inmem]
langchain-google-genai
langchain-mcp-adapters
python-dotenv
gitpython
ipykernel
pydantic